update deal filters
